refactor(repo): разделить схемы и типы пагинации
