+++ packages/types/src/pagination/types/pagination.types.ts
@@ -1,0 +1,16 @@
+export interface PaginationMeta {
+	page: number
+	limit: number
+	total: number
+	totalPages: number
+}
+
+export interface PaginatedResponse<T> {
+	data: T[]
+	meta: PaginationMeta
+}
+
+export interface PaginationParams {
+	page?: number
+	limit?: number
+}
